improve logging

+++ integration/inkscape/conduct_create_setup.py
@@ -4,13 +4,21 @@ import os
 from conduct import conduct
 import subprocess
 
+def log_stub(info):
+    pass
+
+def log(info):
+    inkex.utils.debug(info)
+
 class ConductCreateSetup(inkex.EffectExtension):
 
     def add_arguments(self, pars):
         pars.add_argument("-m", "--manifest", default="", help="Manifest File Path")
 
+
     def effect(self):
         manifest_path = self.options.manifest
+        conduct.log = log_stub
         c = conduct.get_from_manifest_path(manifest_path, "inkscape")
         result = c.setup('.svg')
 
@@ -34,6 +42,11 @@ class ConductCreateSetup(inkex.EffectExtension):
 
         exe = inkex.command.which('inkscape')
 
+        if(exe.startswith('/tmp/.mount')):
+            log("Detected running in an AppImage, this process will hang until the new instance is closed!")
+            inkex.command.inkscape(path)
+            return
+        
         #if we could find a good way to kill the original inkscape instance after starting the new one, that would be ideal
         if os.name == 'nt':
             creation_flags = subprocess.CREATE_NEW_PROCESS_GROUP | subprocess.DETACHED_PROCESS
